+++ docs/changelog.md
@@ -5,6 +5,7 @@
 - Middleware no longer uses environment that may cause side effect
 across sessions.
 - Better instructions for deploying as a service.
+- Remove the deprecated `Logger` class, see [log](https://github.com/devOpifex/log) package.
 
 ## ambiorix 1.0.2